moved add_pending_note to transaction_record
