+++ zingolib/src/wallet/transaction_record.rs
@@ -15,6 +15,7 @@ use zcash_client_backend::{
 };
 use zcash_primitives::{consensus::BlockHeight, transaction::TxId};
 
+use crate::wallet::traits::Recipient as _;
 use crate::{
     error::ZingoLibError,
     wallet::{
@@ -72,6 +73,8 @@ pub struct TransactionRecord {
     pub price: Option<f64>,
 }
 
+// much data assignment of this struct is done through the pub fields as of january 2024. Todo: should have private fields and public methods.
+
 // set
 impl TransactionRecord {
     /// TODO: Add Doc Comment Here!
@@ -110,7 +113,37 @@ impl TransactionRecord {
             }
         }
     }
-    // much data assignment of this struct is done through the pub fields as of january 2024. Todo: should have private fields and public methods.
+    /// adds a note. however, does not fully commit to adding a note, because this note isnt chained into block
+    /// if the transaction is not already recorded, return Err(())
+    pub(crate) fn add_pending_note<D: DomainWalletExt>(
+        &mut self,
+        note: D::Note,
+        to: D::Recipient,
+        output_index: usize,
+    ) {
+        match D::WalletNote::get_record_outputs(self)
+            .iter_mut()
+            .find(|n| n.note() == &note)
+        {
+            None => {
+                let nd = D::WalletNote::from_parts(
+                    to.diversifier(),
+                    note,
+                    None,
+                    None,
+                    None,
+                    None,
+                    // if this is change, we'll mark it later in check_notes_mark_change
+                    false,
+                    false,
+                    Some(output_index as u32),
+                );
+
+                D::WalletNote::transaction_metadata_notes_mut(self).push(nd);
+            }
+            Some(_) => {}
+        }
+    }
 }
 //get
 impl TransactionRecord {
